# custos/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse:
    properties:
      code:
        type: string
      fields:
        additionalProperties: true
        type: object
      message:
        type: string
    type: object
...
  github_com_julesChu12_fly_custos_internal_application_dto.GetProfileResponse:
    properties:
      avatar:
        type: string
      birthday:
        description: 'Format: "2006-01-02"'
        type: string
      email:
        type: string
      extra:
        type: string
      gender:
        type: string
      nickname:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
...
  github_com_julesChu12_fly_custos_internal_application_dto.LoginResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_expires_in:
        type: integer
      refresh_token:
        type: string
      session_id:
        type: string
      token_type:
        type: string
      user:
        $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.UserInfo'
    type: object
  github_com_julesChu12_fly_custos_internal_application_dto.OAuthBindRequest:
    properties:
      code:
        type: string
      redirect_url:
        type: string
      state:
        type: string
    required:
    - code
    - redirect_url
    - state
    type: object
  github_com_julesChu12_fly_custos_internal_application_dto.OAuthBindResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  github_com_julesChu12_fly_custos_internal_application_dto.OAuthBindingInfo:
    properties:
      bound_at:
        type: string
      id:
        type: integer
      provider:
        type: string
      provider_uid:
        type: string
    type: object
  github_com_julesChu12_fly_custos_internal_application_dto.OAuthBindingsResponse:
    properties:
      bindings:
        items:
          $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.OAuthBindingInfo'
        type: array
    type: object
  github_com_julesChu12_fly_custos_internal_application_dto.OAuthUnbindResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
...
  github_com_julesChu12_fly_custos_internal_application_dto.UpdateProfileRequest:
    properties:
      avatar:
        maxLength: 255
        type: string
      birthday:
        description: 'Format: "2006-01-02"'
        type: string
      extra:
        type: string
      gender:
        enum:
        - male
        - female
        - other
        type: string
      nickname:
        maxLength: 64
        type: string
    type: object
  github_com_julesChu12_fly_custos_internal_application_dto.UpdateProfileResponse:
    properties:
      message:
        type: string
      user_id:
        type: integer
    type: object
...
  github_com_julesChu12_fly_custos_internal_application_dto.UserInfo:
    properties:
      avatar:
        type: string
      email:
        type: string
      id:
        type: integer
      nickname:
        type: string
      role:
        type: string
      status:
        type: string
      username:
        type: string
    type: object
...
host: localhost:8081
info:
  contact:
    email: support@fly.com
    name: API Support
    url: http://www.fly.com/support
  description: 用户认证和授权服务API文档，提供用户注册、登录、OAuth、RBAC等功能
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Custos User & Authentication Service API
  version: "1.0"
paths:
...
  /oauth/{provider}/bind:
    post:
      consumes:
      - application/json
      description: 将OAuth提供商账号绑定到当前登录用户
      parameters:
      - description: OAuth提供商（google/github）
        in: path
        name: provider
        required: true
        type: string
      - description: 绑定信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.OAuthBindRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.OAuthBindResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "409":
          description: Conflict
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 绑定OAuth账号
      tags:
      - OAuth
  /oauth/{provider}/callback:
    get:
      description: 处理OAuth提供商的回调并完成登录
      parameters:
      - description: OAuth提供商（google/github）
        in: path
        name: provider
        required: true
        type: string
      - description: 授权码
        in: query
        name: code
        required: true
        type: string
      - description: 状态码
        in: query
        name: state
        required: true
        type: string
      - description: 回调地址
        in: query
        name: redirect_url
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.LoginResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: 处理OAuth回调
      tags:
      - OAuth
  /oauth/{provider}/login:
    get:
      description: 生成指定OAuth提供商的授权URL（支持Google、GitHub）
      parameters:
      - description: OAuth提供商（google/github）
        in: path
        name: provider
        required: true
        type: string
      - description: 回调地址
        in: query
        name: redirect_url
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              auth_url:
                type: string
              state:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
      summary: 获取OAuth授权URL
      tags:
      - OAuth
  /oauth/{provider}/unbind:
    delete:
      description: 将OAuth提供商账号从当前登录用户解绑
      parameters:
      - description: OAuth提供商（google/github）
        in: path
        name: provider
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.OAuthUnbindResponse'
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 解绑OAuth账号
      tags:
      - OAuth
  /oauth/bindings:
    get:
      description: 获取当前登录用户的所有OAuth账号绑定
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.OAuthBindingsResponse'
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      security:
      - BearerAuth: []
      summary: 获取OAuth绑定列表
      tags:
      - OAuth
  /profile:
    get:
      description: 获取当前登录用户的个人资料信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.GetProfileResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 获取个人资料
      tags:
      - 用户配置
    put:
      consumes:
      - application/json
      description: 更新当前登录用户的个人资料信息
      parameters:
      - description: 资料信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.UpdateProfileResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_julesChu12_fly_custos_internal_application_dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: 更新个人资料
      tags:
      - 用户配置
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...